refactor: move flat implementations into separate module
