+++ crates/bitwarden-error-macro/src/lib.rs
@@ -1,8 +1,8 @@
 mod basic;
+mod flat;
 
 use darling::{ast::NestedMeta, FromMeta};
 use quote::quote;
-use syn::Data;
 
 #[derive(FromMeta)]
 struct BitwardenErrorArgs {
@@ -47,21 +47,9 @@ pub fn bitwarden_error(
     let input = syn::parse_macro_input!(item as syn::DeriveInput);
     match error_type {
         BitwardenErrorType::Basic => basic::attribute::bitwarden_error_basic(&input),
-        BitwardenErrorType::Flat => flat_error_impl(&input),
+        BitwardenErrorType::Flat => flat::attribute::bitwarden_error_flat(&input),
         BitwardenErrorType::Full => full_error_impl(&input),
     }
-}
-
-fn flat_error_impl(input: &syn::DeriveInput) -> proc_macro::TokenStream {
-    let type_identifier = &input.ident;
-
-    quote! {
-        #[derive(FlatError)]
-        #input
-
-        impl BitwardenError for #type_identifier {}
-    }
-    .into()
 }
 
 fn full_error_impl(input: &syn::DeriveInput) -> proc_macro::TokenStream {
@@ -87,118 +75,5 @@ pub fn basic_error(item: proc_macro::TokenStream) -> proc_macro::TokenStream {
 
 #[proc_macro_derive(FlatError)]
 pub fn flat_error(item: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    let input = syn::parse_macro_input!(item as syn::DeriveInput);
-    let type_identifier = &input.ident;
-
-    match &input.data {
-        Data::Enum(data) => {
-            let variant_names = data.variants.iter().map(|variant| &variant.ident);
-
-            let match_arms = data.variants.iter().map(|variant| match variant.fields {
-                syn::Fields::Unit => {
-                    let variant_ident = &variant.ident;
-                    let variant_name = format!("{}", variant_ident);
-                    quote! {
-                        #type_identifier::#variant_ident => #variant_name
-                    }
-                }
-                syn::Fields::Named(_) => {
-                    let variant_ident = &variant.ident;
-                    let variant_name = format!("{}", variant_ident);
-                    quote! {
-                        #type_identifier::#variant_ident { .. } => #variant_name
-                    }
-                }
-                syn::Fields::Unnamed(_) => {
-                    let variant_ident = &variant.ident;
-                    let variant_name = format!("{}", variant_ident);
-                    quote! {
-                        #type_identifier::#variant_ident(..) => #variant_name
-                    }
-                }
-            });
-
-            let wasm = flat_error_wasm(&type_identifier, &variant_names.collect::<Vec<_>>());
-
-            quote! {
-                #wasm
-
-                #[automatically_derived]
-                impl FlatError for #type_identifier {
-                    fn error_variant(&self) -> &'static str {
-                        match &self {
-                            #(#match_arms), *
-                        }
-                    }
-                }
-            }
-            .into()
-        }
-        Data::Struct(_) => {
-            let type_identifier = &input.ident;
-            let variant_name = format!("{}", type_identifier);
-
-            quote! {
-                #[automatically_derived]
-                impl FlatError for #type_identifier {
-                    fn error_variant(&self) -> &'static str {
-                        #variant_name
-                    }
-                }
-            }
-            .into()
-        }
-        Data::Union(_) => {
-            syn::Error::new_spanned(input, "bitwarden_error cannot be used with unions")
-                .to_compile_error()
-                .into()
-        }
-    }
-}
-
-#[cfg(feature = "wasm")]
-fn flat_error_wasm(
-    type_identifier: &proc_macro2::Ident,
-    variant_names: &[&proc_macro2::Ident],
-) -> proc_macro2::TokenStream {
-    let ts_identifier = quote::format_ident!("TS_TYPES_{}", type_identifier);
-    let ts_code_str = format!(
-        r##"r#"
-            export interface {} extends Error {{
-                name: "{}";
-                variant: {};
-            }};
-        "#"##,
-        type_identifier,
-        type_identifier,
-        variant_names
-            .iter()
-            .map(|vn| format!(r#""{vn}""#))
-            .collect::<Vec<String>>()
-            .join("|")
-    );
-    let ts_code: proc_macro2::TokenStream = ts_code_str.parse().unwrap();
-
-    quote! {
-        #[wasm_bindgen(typescript_custom_section)]
-        const #ts_identifier: &'static str = #ts_code;
-
-        #[automatically_derived]
-        impl From<#type_identifier> for JsValue {
-            fn from(error: #type_identifier) -> Self {
-                let js_error = JsError::new(error.to_string());
-                js_error.set_name(stringify!(#type_identifier).to_owned());
-                js_error.set_variant(error.error_variant().to_owned());
-                js_error.into()
-            }
-        }
-    }
-}
-
-#[cfg(not(feature = "wasm"))]
-fn flat_error_wasm(
-    type_identifier: &proc_macro2::Ident,
-    variant_names: &[&proc_macro2::Ident],
-) -> proc_macro2::TokenStream {
-    quote! {}
+    flat::derive::flat_error(item)
 }
